# Optimize problemset PDF transcription task