--- a/pyproject.toml
+++ b/pyproject.toml
@@ -12,7 +12,7 @@ requires-poetry = ">2.1,<3"
 
 
 [tool.poetry.dependencies]
-python = "~3.12"
+python = ">3.12,<3.13"
 Django = "4.2.23"
 attrs = "^25.0.0"
 base36 = "^0.1.1"
@@ -111,7 +111,7 @@ uwsgi = "^2.0.29"
 uwsgitop = "^0.12"
 wrapt = "^1.14.1"
 youtube-transcript-api = "^1.0.0"
-pymupdf = "^1.26.3"
+pymupdf = "1.23.5"
 
 [tool.poetry.group.dev.dependencies]
 bpython = "^0.25"