--- a/pyproject.toml
+++ b/pyproject.toml
@@ -12,7 +12,7 @@ requires-poetry = ">2.1,<3"
 
 
 [tool.poetry.dependencies]
-python = "^3.12.0,<3.13"
+python = "~3.12"
 Django = "4.2.23"
 attrs = "^25.0.0"
 base36 = "^0.1.1"