--- a/pyproject.toml
+++ b/pyproject.toml
@@ -111,7 +111,7 @@ uwsgi = "^2.0.29"
 uwsgitop = "^0.12"
 wrapt = "^1.14.1"
 youtube-transcript-api = "^1.0.0"
-pymupdf = "1.23.5"
+pymupdf = "1.23.6"
 
 [tool.poetry.group.dev.dependencies]
 bpython = "^0.25"