--- a/pyproject.toml
+++ b/pyproject.toml
@@ -12,7 +12,7 @@ requires-poetry = ">2.1,<3"
 
 
 [tool.poetry.dependencies]
-python = ">=3.10,<3.13"
+python = "~3.12"
 Django = "4.2.23"
 attrs = "^25.0.0"
 base36 = "^0.1.1"
@@ -50,7 +50,7 @@ feedparser = "^6.0.10"
 google-api-python-client = "^2.89.0"
 html2text = "^2025.0.0"
 html5lib = "^1.1"
-ipython = "^9.0.0"
+ipython = { version = "^9.0.0", markers = "python_version >= '3.11'" }
 isodate = "^0.7.2"
 jedi = "^0.19.0"
 langchain = "^0.3.11"