--- a/pyproject.toml
+++ b/pyproject.toml
@@ -1,8 +1,3 @@
-[project]
-requires-python = ">=3.12"
-name = "mit-learn"
-
-
 [tool.poetry]
 # Don't try to install mit-learn as a package on deploy
 package-mode = false
@@ -14,7 +9,6 @@ readme = "README.md"
 packages = []
 authors  = ["MIT ODL"]
 requires-poetry = ">2.1,<3"
-
 
 
 [tool.poetry.dependencies]